[TM-43] fix: Fixed Http service flaws
