+++ src/MoySkladApi/MoySkladClient/index.ts
@@ -1,5 +1,5 @@
 import { RetryService } from '../MoySkladServices/RetryService';
-import { HTTPService } from '../MoySkladServices/HTTPService';
+import { HttpService } from '../MoySkladServices/HTTPService';
 import { getObjectInLowercase } from '../../utils/objects';
 import { sleep } from '../../utils/constants';
 import {
@@ -29,7 +29,7 @@ export class MoySkladClient {
         path: TRelativePath,
         params?: Record<string, any>
     ): string {
-        const querySet = HTTPService.buildQuery(params);
+        const querySet = HttpService.buildQuery(params);
 
         const url = `${baseURL}/${path}`;
 
@@ -119,7 +119,7 @@ export class MoySkladClient {
                 statusText,
                 headers: respHeaders,
                 data,
-            } = await HTTPService.requestWithJson<T>({
+            } = await HttpService.requestWithJson<T>({
                 url,
                 method,
                 headers,
@@ -193,7 +193,7 @@ export class MoySkladClient {
                     return { status, headers: respHeaders, data: data as T };
                 }
 
-                const message = HTTPService.pickErrorMessage(
+                const message = HttpService.pickErrorMessage(
                     data,
                     statusText ?? ''
                 );
